+++ packages/docusaurus-theme-openapi-docs/src/theme/ApiDemoPanel/ParamOptions/index.tsx
@@ -5,19 +5,17 @@
  * LICENSE file in the root directory of this source tree.
  * ========================================================================== */
 
-// @ts-nocheck
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 
-import { ErrorMessage } from "@hookform/error-message";
-import { nanoid } from "@reduxjs/toolkit";
 import FormItem from "@theme/ApiDemoPanel/FormItem";
-import FormMultiSelect from "@theme/ApiDemoPanel/FormMultiSelect";
-import FormSelect from "@theme/ApiDemoPanel/FormSelect";
-import FormTextInput from "@theme/ApiDemoPanel/FormTextInput";
-import { useTypedDispatch, useTypedSelector } from "@theme/ApiItem/hooks";
-import { Controller, useFormContext } from "react-hook-form";
+import ParamArrayFormItem from "@theme/ApiDemoPanel/ParamOptions/ParamFormItems/ParamArrayFormItem";
+import ParamBooleanFormItem from "@theme/ApiDemoPanel/ParamOptions/ParamFormItems/ParamBooleanFormItem";
+import ParamMultiSelectFormItem from "@theme/ApiDemoPanel/ParamOptions/ParamFormItems/ParamMultiSelectFormItem";
+import ParamSelectFormItem from "@theme/ApiDemoPanel/ParamOptions/ParamFormItems/ParamSelectFormItem";
+import ParamTextFormItem from "@theme/ApiDemoPanel/ParamOptions/ParamFormItems/ParamTextFormItem";
+import { useTypedSelector } from "@theme/ApiItem/hooks";
 
-import { Param, setParam } from "./slice";
+import { Param } from "./slice";
 
 export interface ParamProps {
   param: Param;
@@ -141,241 +139,6 @@ function ParamOptions() {
         </>
       )}
     </>
-  );
-}
-
-function ArrayItem({
-  param,
-  onChange,
-}: ParamProps & { onChange(value?: string): any }) {
-  if (param.schema?.items?.type === "boolean") {
-    return (
-      <FormSelect
-        options={["---", "true", "false"]}
-        onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
-          const val = e.target.value;
-          onChange(val === "---" ? undefined : val);
-        }}
-      />
-    );
-  }
-
-  return (
-    <FormTextInput
-      // {...register(`${param.name}`)}
-      placeholder={param.description || param.name}
-      onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
-        onChange(e.target.value);
-      }}
-    />
-  );
-}
-
-function ParamArrayFormItem({ param }: ParamProps) {
-  const [items, setItems] = useState<{ id: string; value?: string }[]>([]);
-  const dispatch = useTypedDispatch();
-
-  function handleAddItem() {
-    setItems((i) => [
-      ...i,
-      {
-        id: nanoid(),
-      },
-    ]);
-  }
-
-  useEffect(() => {
-    const values = items
-      .map((item) => item.value)
-      .filter((item): item is string => !!item);
-
-    dispatch(
-      setParam({
-        ...param,
-        value: values.length > 0 ? values : undefined,
-      })
-    );
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [items]);
-
-  function handleDeleteItem(itemToDelete: { id: string }) {
-    return () => {
-      const newItems = items.filter((i) => i.id !== itemToDelete.id);
-      setItems(newItems);
-    };
-  }
-
-  function handleChangeItem(itemToUpdate: { id: string }) {
-    return (value: string) => {
-      const newItems = items.map((i) => {
-        if (i.id === itemToUpdate.id) {
-          return { ...i, value: value };
-        }
-        return i;
-      });
-      setItems(newItems);
-    };
-  }
-
-  return (
-    <>
-      {items.map((item) => (
-        <div key={item.id} style={{ display: "flex" }}>
-          <ArrayItem param={param} onChange={handleChangeItem(item)} />
-          <button
-            className="openapi-demo__delete-btn"
-            onClick={handleDeleteItem(item)}
-          >
-            <svg
-              focusable="false"
-              preserveAspectRatio="xMidYMid meet"
-              xmlns="http://www.w3.org/2000/svg"
-              fill="currentColor"
-              width="16"
-              height="16"
-              viewBox="0 0 32 32"
-              aria-hidden="true"
-            >
-              <path d="M24 9.4L22.6 8 16 14.6 9.4 8 8 9.4 14.6 16 8 22.6 9.4 24 16 17.4 22.6 24 24 22.6 17.4 16 24 9.4z"></path>
-              <title>Delete</title>
-            </svg>
-          </button>
-        </div>
-      ))}
-      <button className="openapi-demo__thin-btn" onClick={handleAddItem}>
-        Add item
-      </button>
-    </>
-  );
-}
-
-function ParamSelectFormItem({ param }: ParamProps) {
-  const dispatch = useTypedDispatch();
-
-  const options = param.schema?.enum ?? [];
-
-  return (
-    <FormSelect
-      options={["---", ...(options as string[])]}
-      onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
-        const val = e.target.value;
-        dispatch(
-          setParam({
-            ...param,
-            value: val === "---" ? undefined : val,
-          })
-        );
-      }}
-    />
-  );
-}
-
-function ParamBooleanFormItem({ param }: ParamProps) {
-  const dispatch = useTypedDispatch();
-
-  return (
-    <FormSelect
-      options={["---", "true", "false"]}
-      onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
-        const val = e.target.value;
-        dispatch(
-          setParam({
-            ...param,
-            value: val === "---" ? undefined : val,
-          })
-        );
-      }}
-    />
-  );
-}
-
-function ParamMultiSelectFormItem({ param }: ParamProps) {
-  const {
-    control,
-    formState: { errors },
-  } = useFormContext();
-
-  const showErrorMessage = errors?.paramMultiSelect;
-
-  const dispatch = useTypedDispatch();
-
-  const options = param.schema?.items?.enum ?? [];
-
-  const pathParams = useTypedSelector((state: any) => state.params.path);
-  const queryParams = useTypedSelector((state: any) => state.params.query);
-  const cookieParams = useTypedSelector((state: any) => state.params.cookie);
-  const headerParams = useTypedSelector((state: any) => state.params.header);
-
-  const paramTypeToWatch = pathParams.length
-    ? pathParams
-    : queryParams.length
-    ? queryParams
-    : cookieParams.length
-    ? cookieParams
-    : headerParams;
-
-  const handleChange = (e: any, onChange: any) => {
-    const values = Array.prototype.filter
-      .call(e.target.options, (o) => o.selected)
-      .map((o) => o.value);
-
-    dispatch(
-      setParam({
-        ...param,
-        value: values.length > 0 ? values : undefined,
-      })
-    );
-
-    onChange(paramTypeToWatch);
-  };
-
-  return (
-    <>
-      <Controller
-        control={control}
-        rules={{ required: param.required ? "This field is required" : false }}
-        name="paramMultiSelect"
-        render={({ field: { onChange, name } }) => (
-          <FormMultiSelect
-            options={options as string[]}
-            name={name}
-            onChange={(e: any) => handleChange(e, onChange)}
-            showErrors={showErrorMessage}
-          />
-        )}
-      />
-      {showErrorMessage && (
-        <ErrorMessage
-          errors={errors}
-          name="paramMultiSelect"
-          render={({ message }) => (
-            <div className="openapi-demo__input-error">{message}</div>
-          )}
-        />
-      )}
-    </>
-  );
-}
-
-function ParamTextFormItem({ param }: ParamProps) {
-  const dispatch = useTypedDispatch();
-  return (
-    <FormTextInput
-      isRequired={param.required}
-      paramName={param.name}
-      placeholder={param.description || param.name}
-      onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-        dispatch(
-          setParam({
-            ...param,
-            value:
-              param.in === "path" || param.in === "query"
-                ? e.target.value.replace(/\s/g, "%20")
-                : e.target.value,
-          })
-        )
-      }
-    />
   );
 }
